Move localized and localized_draft to TranslatableMixin
