+++ wagtail/models/__init__.py
@@ -1070,42 +1070,6 @@ class Page(AbstractPage, index.Indexed, ClusterableModel, metaclass=PageBase):
         else:
             return self.specific_class.get_verbose_name()
 
-    @property
-    def localized_draft(self):
-        """
-        Finds the translation in the current active language.
-
-        If there is no translation in the active language, self is returned.
-
-        Note: This will return translations that are in draft. If you want to exclude
-        these, use the ``.localized`` attribute.
-        """
-        try:
-            locale = Locale.get_active()
-        except (LookupError, Locale.DoesNotExist):
-            return self
-
-        if locale.id == self.locale_id:
-            return self
-
-        return self.get_translation_or_none(locale) or self
-
-    @property
-    def localized(self):
-        """
-        Finds the translation in the current active language.
-
-        If there is no translation in the active language, self is returned.
-
-        Note: This will not return the translation if it is in draft.
-        If you want to include drafts, use the ``.localized_draft`` attribute instead.
-        """
-        localized = self.localized_draft
-        if not localized.live:
-            return self
-
-        return localized
-
     def route(self, request, path_components):
         if path_components:
             # request is for a child of this page
